move query test to integration folder, improve dummy data
