+++ lib/dai-plugin-mcd/test/CdpManager.spec.js
@@ -53,16 +53,16 @@ test('getCdp looks up ilk', async () => {
 
 test('getCombinedEventHistory', async () => {
   const proxy = await maker.currentProxy();
-  const mockFn = jest.fn(async () => dummyEventData);
+  const mockFn = jest.fn(async () => dummyEventData('ETH-A'));
   maker.service(
     ServiceRoles.QUERY_API
   ).getCdpEventsForArrayOfIlksAndUrns = mockFn;
   const events = await cdpMgr.getCombinedEventHistory(proxy);
   expect(mockFn).toBeCalled();
-  const currency = maker
+  const GEM = maker
     .service(ServiceRoles.CDP_TYPE)
     .getCdpType(null, events[0].ilk).currency;
-  expect(events).toEqual(formattedDummyEventData(currency));
+  expect(events).toEqual(formattedDummyEventData(GEM, events[0].ilk));
 });
 
 test('transaction tracking for openLockAndDraw', async () => {
