+++ lib/dai-plugin-mcd/test/ManagedCdp.spec.js
@@ -1,4 +1,5 @@
 import { mcdMaker, setupCollateral, setDebtCeiling } from './helpers';
+import { stringToBytes } from '../src/utils';
 import { ETH, COL1, MDAI } from '../src';
 import { ServiceRoles } from '../src/constants';
 
@@ -6,25 +7,28 @@ const { CDP_MANAGER, QUERY_API } = ServiceRoles;
 
 let maker, dai, proxy, txMgr;
 
-export const dummyEventData = [
-  {
-    dart: '-6000000000000000000',
-    dink: '0',
-    ilk: { nodes: [{ rate: '2000000000000000000000000000' }] },
-    tx: {
-      transactionHash:
-        '0xbe023a205453b833e65bf29063de8b8b3bd44d2e68c9c079f681ec46a765a63f',
-      txFrom: '0x1ad35418e7b7c5746ea42295a1100480a810256a',
-      era: { iso: '1970-01-01T00:01:20' }
-    },
-    urn: {
-      nodes: [{ art: '6849191748702380000', ink: '5336190215977040000' }]
-    },
-    ilkName: '0x4554482d41'
-  }
-];
+export function dummyEventData(ilk) {
+  const ilkName = stringToBytes(ilk);
+  return [
+    {
+      dart: '-6000000000000000000',
+      dink: '0',
+      ilk: { nodes: [{ rate: '2000000000000000000000000000' }] },
+      tx: {
+        transactionHash:
+          '0xbe023a205453b833e65bf29063de8b8b3bd44d2e68c9c079f681ec46a765a63f',
+        txFrom: '0x1ad35418e7b7c5746ea42295a1100480a810256a',
+        era: { iso: '1970-01-01T00:01:20' }
+      },
+      urn: {
+        nodes: [{ art: '6849191748702380000', ink: '5336190215977040000' }]
+      },
+      ilkName
+    }
+  ];
+}
 
-export function formattedDummyEventData(GEM) {
+export function formattedDummyEventData(GEM, ilk) {
   return [
     {
       transactionHash:
@@ -36,7 +40,7 @@ export function formattedDummyEventData(GEM) {
       senderAddress: '0x1ad35418e7b7c5746ea42295a1100480a810256a',
       resultingCollateral: GEM(5.33619021597704),
       resultingDebt: MDAI(6.84919174870238),
-      ilk: 'ETH-A'
+      ilk
     }
   ];
 }
@@ -167,15 +171,12 @@ describe.each([
   });
 
   test('getEventHistory', async () => {
-    const mockFn = jest.fn(async () => dummyEventData);
+    const mockFn = jest.fn(async () => dummyEventData(ilk));
     maker.service(QUERY_API).getCdpEventsForIlkAndUrn = mockFn;
     const cdp = await maker.service(CDP_MANAGER).open(ilk);
     const events = await cdp.getEventHistory();
     expect(mockFn).toBeCalled();
-    const currency = maker
-      .service(ServiceRoles.CDP_TYPE)
-      .getCdpType(null, events[0].ilk).currency;
-    expect(events).toEqual(formattedDummyEventData(currency));
+    expect(events).toEqual(formattedDummyEventData(GEM, ilk));
   });
 
   test('openLock, lock, lockAndDraw, free', async () => {
